requirements.txt
# GCP Libraries
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.10.0
google-cloud-pipeline-components>=2.0.0
google-cloud-dataflow==2.4.0

# ML Libraries
tensorflow>=2.13.0,<2.16.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
xgboost>=1.7.0

# MLOps Tools
mlflow>=2.8.0
kubeflow-pipelines>=2.0.0
tfx>=1.14.0

# Development Tools
jupyter>=1.0.0
notebook>=7.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.5.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
pyyaml>=6.0.0
tqdm>=4.65.0

# Additional ML tools
lightgbm>=4.0.0
catboost>=1.2.0